requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==24.2.0
awsebcli==3.25.2
bcrypt==5.0.0
beautifulsoup4==4.12.3
billiard==4.2.1
blessed==1.23.0
botocore==1.40.69
bs4==0.0.2
celery==5.4.0
cement==2.10.14
certifi==2024.8.30
cffi==2.0.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==46.0.3
decorator==5.2.1
Deprecated==1.3.1
Django==5.1.2
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
fabric==3.2.2
gunicorn==23.0.0
h11==0.14.0
idna==3.8
invoke==2.2.1
jmespath==1.0.1
kombu==5.4.2
outcome==1.3.0.post0
packaging==24.2
paramiko==4.0.0
pathspec==0.12.1
prompt_toolkit==3.0.48
psycopg==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
PyNaCl==1.6.0
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.3
requests==2.32.3
selenium==4.24.0
semantic-version==2.10.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
termcolor==2.5.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wrapt==2.0.1
wsproto==1.2.0
